requirements.txt
dj-database-url==2.2.0
Django==5.2.1
django-debug-toolbar==5.2.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
environs==14.1.1
geopy==2.4.1
phonenumbers==9.0.5
Pillow==8.2.0
requests==2.32.3
